+++ capio/server/include/remote/backend/none.hpp
@@ -1,15 +1,15 @@
-#ifndef CAPIO_DEFAULT_HPP
-#define CAPIO_DEFAULT_HPP
+#ifndef CAPIO_SERVER_REMOTE_BACKEND_NONE_HPP
+#define CAPIO_SERVER_REMOTE_BACKEND_NONE_HPP
 #include "remote/backend.hpp"
 
-class NoBackend final : public Backend {
+class NoneBackend final : public Backend {
   public:
-    NoBackend(int argc, char **argv);
-    ~NoBackend() override = default;
+    NoneBackend(int argc, char **argv);
+    ~NoneBackend() override = default;
     void handshake_servers() override;
     RemoteRequest read_next_request() override;
     void send_file(char *shm, long int nbytes, const std::string &target) override;
     void send_request(const char *message, int message_len, const std::string &target) override;
     void recv_file(char *shm, const std::string &source, long int bytes_expected) override;
 };
-#endif // CAPIO_DEFAULT_HPP
+#endif // CAPIO_SERVER_REMOTE_BACKEND_NONE_HPP
